feat(B1): POST /files (F2-04) + factory in domain and null-safe validation
