+++ api-service/src/main/java/com/practice/apiservice/dto/FileUploadResponse.java
@@ -1,0 +1,5 @@
+package com.practice.apiservice.dto;
+
+import java.util.UUID;
+
+public record FileUploadResponse(UUID id, String originalFilename) {}
