+++ api-service/src/main/java/com/practice/apiservice/exception/ApiErrorHandler.java
@@ -1,0 +1,52 @@
+package com.practice.apiservice.exception;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import com.practice.apiservice.utils.error.FileTooLargeException;
+
+@RestControllerAdvice
+public class ApiErrorHandler {
+    record FieldItem(String field, String message) {}
+    record ApiError(String code, String message, List<FieldItem> fields) {}
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    ResponseEntity<ApiError> onValidation(MethodArgumentNotValidException ex) {
+        var fields = ex.getBindingResult().getFieldErrors().stream()
+                .map(fe -> new FieldItem(fe.getField(), fe.getDefaultMessage()))
+                .toList();
+        return ResponseEntity.badRequest()
+                .body(new ApiError("VALIDATION_ERROR", "Payload inválido", fields));
+    }
+
+    @ExceptionHandler(FileTooLargeException.class)
+    ResponseEntity<ApiError> onTooLarge(FileTooLargeException ex) {
+        return ResponseEntity.badRequest()
+                .body(new ApiError("FILE_TOO_LARGE",
+                        "Max " + ex.getMax() + " bytes, recibió " + ex.getActual(), List.of()));
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    ResponseEntity<ApiError> onIllegalArgument(IllegalArgumentException ex) {
+        List<FieldItem> fields = List.of();
+        
+        String message = ex.getMessage();
+        if (message != null) {
+            if (message.contains("originalFilename is blank")) {
+                fields = List.of(new FieldItem("originalFilename", "must not be blank"));
+            } else if (message.contains("checksumSha256 must contain exactly 64 hexadecimal characters")) {
+                fields = List.of(new FieldItem("checksumSha256", "must be a valid SHA-256 hash"));
+            } else if (message.contains("storagePath is blank")) {
+                fields = List.of(new FieldItem("storagePath", "must not be blank"));
+            } else if (message.contains("sizeBytes must be > 0")) {
+                fields = List.of(new FieldItem("sizeBytes", "must be greater than 0"));
+            }
+        }
+
+        return ResponseEntity.badRequest()
+                .body(new ApiError("VALIDATION_ERROR", "Validation failed", fields));
+    }
+}
