+++ core/src/test/java/com/AppTest.java
@@ -1,19 +1,5 @@
 package com;
 
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void testApp()
-    {
-        assertTrue( true );
-    }
+public class AppTest {
+    
 }
