+++ core/src/test/java/com/practice/AppTest.java
@@ -1,0 +1,19 @@
+package com.practice;
+
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void testApp()
+    {
+        assertTrue( true );
+    }
+}
