+++ core/src/test/java/com/practice/BatchJobConfigTest.java
@@ -1,4 +1,4 @@
-package com;
+package com.practice;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
