+++ core/src/test/java/com/practice/ErrorHandlerTest.java
@@ -1,4 +1,4 @@
-package com;
+package com.practice;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertTrue;
